Add files via upload
Marketing campaigns play a crucial role in promoting products or services to potential customers. However, designing effective campaigns requires an understanding of customer behavior and preferences. Predictive modeling offers a valuable tool for marketers to anticipate customer responses and tailor their strategies accordingly. In this project, we aim to develop a neural network-based predictive model to forecast customer responses to marketing campaigns. The dataset contains a range of features, including demographic information and past interaction history, which will be leveraged to train the model.